Server/requirements.txt
absl-py==2.1.0
anyio==4.2.0
APScheduler==3.10.1
asgiref==3.3.1
astroid==2.4.2
asttokens==2.4.1
backcall==0.2.0
blinker==1.7.0
boto3==1.16.47
botocore==1.19.47
cachetools==5.3.2
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.0
decorator==5.1.1
defusedxml==0.7.1
Django==3.1.5
django-extensions==3.1.0
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.2
docker==6.1.3
executing==2.0.1
fastapi==0.96.0
filelock==3.13.1
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.47.2
fsspec==2024.2.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
GPUtil==1.4.0
grpcio==1.60.1
gunicorn==20.0.4
humanfriendly==10.0
idna==2.10
imageio==2.33.1
inference==0.9.8
iniconfig==2.0.0
ipython==8.21.0
isort==5.6.4
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==0.10.0
kiwisolver==1.4.5
lazy-object-proxy==1.4.3
lazy_loader==0.3
Markdown==3.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.6.1
mdurl==0.1.2
mpmath==1.3.0
mysqlclient==2.2.3
networkx==3.2.1
numpy==1.23.5
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.15.1
onnxsim==0.4.35
opencv-python==4.8.0.76
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
piexif==1.1.3
Pillow==8.0.1
pluggy==1.4.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.0.0
prompt-toolkit==3.0.43
protobuf==4.21.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybase64==1.3.2
pydantic==1.10.14
pydotplus==2.0.2
Pygments==2.17.2
PyJWT==2.0.0
pylint==2.6.0
pyparsing==2.4.7
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.6.0
pyqt6-tools==6.4.2.3.3
pytest==8.0.0
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2020.5
PyYAML==6.0.1
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
redis==5.0.1
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.5.2
rsa==4.9
s3transfer==0.3.3
scikit-image==0.22.0
scipy==1.12.0
seaborn==0.13.2
shapely==2.0.1
six==1.15.0
sniffio==1.3.0
sqlparse==0.4.1
stack-data==0.6.3
starlette==0.27.0
supervision==0.18.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
tifffile==2024.1.30
toml==0.10.2
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
traitlets==5.14.1
twilio==6.50.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==1.26.2
wcwidth==0.2.13
websocket-client==1.7.0
Werkzeug==3.0.1
whitenoise==5.2.0
wrapt==1.12.1
